Clean up and refactor RDO functions and data
Move RDO decision functions to rdo.rs, extract rd_cost computation function, move some data to appropriate locations.
