+++ src/partition.rs
@@ -147,6 +147,19 @@ pub enum PredictionMode {
     NEW_NEWMV,
 }
 
+pub static RAV1E_PARTITION_TYPES: &'static [PartitionType] = &[
+    PartitionType::PARTITION_NONE,
+    PartitionType::PARTITION_SPLIT
+];
+
+pub static RAV1E_INTRA_TX_TYPES: &'static [TxType] = &[
+    TxType::DCT_DCT,
+    TxType::ADST_DCT,
+    TxType::DCT_ADST,
+    TxType::ADST_ADST,
+    TxType::IDTX
+];
+
 use plane::*;
 use predict::*;
 use context::*;
